+++ demosys/scene/mesh.py
@@ -33,8 +33,8 @@ class Mesh:
             self.mesh_shader.draw(self, proj_mat, view_mat)
 
     def draw_bbox(self, proj_matrix, view_matrix, shader, vao):
-        shader.uniform("m_proj", proj_matrix.astype('f4').tobytes())
-        shader.uniform("m_mv", view_matrix.astype('f4').tobytes())
+        shader.uniform("m_proj", proj_matrix.tobytes())
+        shader.uniform("m_mv", view_matrix.tobytes())
         shader.uniform("bb_min", self.bbox_min.astype('f4').tobytes())
         shader.uniform("bb_max", self.bbox_max.astype('f4').tobytes())
         shader.uniform("color", (0.75, 0.75, 0.75))
